Remove and rename files
