--- include/model.hpp
@@ -1,14 +1,0 @@
-//
-//  model.hpp
-//  compnal
-//
-//  Created by Kohei Suzuki on 2021/05/23.
-//
-
-#ifndef model_hpp
-#define model_hpp
-
-#include "../src/model/heisenberg_model_1d.hpp"
-#include "../src/model/model_utility.hpp"
-
-#endif /* model_hpp */
