--- include/solver.hpp
@@ -1,13 +1,0 @@
-//
-//  solver.hpp
-//  compnal
-//
-//  Created by Kohei Suzuki on 2021/05/23.
-//
-
-#ifndef solver_hpp
-#define solver_hpp
-
-#include "../src/solver/exact_diag_1d.hpp"
-
-#endif /* solver_hpp */
