--- include/sparse_matrix.hpp
@@ -1,16 +1,0 @@
-//
-//  sparse_matrix.hpp
-//  compnal
-//
-//  Created by Kohei Suzuki on 2021/05/22.
-//
-
-#ifndef sparse_matrix_hpp
-#define sparse_matrix_hpp
-
-#include "../src/sparse_matrix/compressed_row_storage.hpp"
-#include "../src/sparse_matrix/braket_vector.hpp"
-#include "../src/sparse_matrix/lapack.hpp"
-#include "../src/sparse_matrix/eigendecomposition.hpp"
-
-#endif /* sparse_matrix_hpp */
