+++ src/sparse_matrix/compressed_row_storage.hpp
@@ -8,7 +8,7 @@
 #ifndef COMPNAL_SPARSE_MATRIX_COMPRESSED_ROW_STORAGE_HPP_
 #define COMPNAL_SPARSE_MATRIX_COMPRESSED_ROW_STORAGE_HPP_
 
-#include "../utility/utility.hpp"
+#include "../utility/all.hpp"
 #include <iostream>
 #include <cstdint>
 #include <vector>
@@ -125,7 +125,7 @@ struct CRS {
    }
    
    void SortCol() {
-#pragma omp parallel for schedule(guided)
+#pragma omp parallel for schedule (guided)
       for (std::size_t i = 0; i < this->row_dim; ++i) {
          utility::QuickSort<std::size_t, RealType>(&this->col, &this->val, this->row[i], this->row[i + 1]);
       }
