+++ src/utility/all.hpp
@@ -1,0 +1,14 @@
+//
+//  all.hpp
+//  compnal
+//
+//  Created by Kohei Suzuki on 2021/11/06.
+//
+
+#ifndef COMPNAL_UTILITY_ALL_HPP_
+#define COMPNAL_UTILITY_ALL_HPP_
+
+#include "model.hpp"
+#include "sort.hpp"
+
+#endif /* COMPNAL_UTILITY_ALL_HPP_ */
