+++ tests/test_main.cpp
@@ -5,8 +5,6 @@
 //  Created by Kohei Suzuki on 2021/06/14.
 //
 
-//#include "test_exact_diag.hpp"
-//#include "test_model.hpp"
 #include "sparse_matrix/test_compressed_row_storage.hpp"
 #include "sparse_matrix/test_eigendecomposition.hpp"
 #include "sparse_matrix/test_braket_vector.hpp"
